Sonarqube issues
Minor:

Make the type of this variable a reference-to-const.

STL algorithms and range-based for loops should be preferred to
traditional for loops.

Critical:

Default capture should not be used.

If the lambda is used immediately, there is no risk of dangling
reference.

This rule raises an issue when default capture is used, unless the
lambda is immediately executed.