Sonarqube issues
Make the type of this variable a reference-to-const.

Switch statements should have 'default' a clause.

Member variables should not be protected, but should be private.

When an object provides encapsulation by using protected member
variables, data integrity logic can be spread through the class and all
its derived class, becoming a source of complexity and that will be
error-prone for maintainers and extenders.

That’s why protected member variables should be changed to private and manipulated exclusively through public or protected member functions of the base class.

Collapsible 'if' statements should be merged.

'Switch' statements should have 'default' clauses.

Empty 'case' clauses that fall through to the 'default' should be omitted.