Sonarqube issues
'if','switch', and range-based for loop initializer should be used to
reduce scope of variables

Redundant access specifiers should be removed because they needlessly
clutter the code.

Empty 'case' clauses that fall through to the 'default' should be
omitted

Special member function should not be defined unless a non standard
behavior is required.

It is good practice to organize your code so that you can use the
default version

'auto' should be used to avoid repetition of types.

When used as a type specifier in a declaration, auto allows the
compiler to deduce the type of a variable based on the type of the
initialization expression.