SDL support (enables rpi4 and M1)
Rename addEvent to addFdEvent

With the addition of addTimerEvent, the naming of addEvent is specific
to the management of an fd, while the naming is generic.

Update the name to make the naming scheme consistent in specifying the
type of event to be added.

Add timer events to event loop.

Extend the EventLoop class to support periodic timer events. This can be
used to run tasks periodically, such as handling the event loop of SDL.

Add SDL sink with initial YUYV support.

Add MJPG support to SDL sink.

Pixel format option -p, added to enable selection of formats.